Move `data-list/HasHiddenFields.js` to `publish/HasHiddenFields.js`.
